requirements.txt
numpy==1.26.4
opencv_contrib_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==10.3.0
streamlit==1.34.0
tensorflow==2.15.0
tensorflow_intel==2.15.0
